fix: adjust test organization
